Applying PR feedback

+++ onnx/backend/test/case/node/split.py
@@ -14,7 +14,7 @@ class Split(Base):
 
     @staticmethod
     def export_1d():
-        input = np.array([1., 2., 3., 4., 5., 6.])
+        input = np.array([1., 2., 3., 4., 5., 6.]).astype(np.float32)
 
         node = onnx.helper.make_node(
             'Split',
@@ -23,7 +23,7 @@ class Split(Base):
             axis=0
         )
 
-        expected_outputs = [np.array([1, 2]), np.array([3, 4]), np.array([5, 6])]
+        expected_outputs = [np.array([1., 2.]).astype(np.float32), np.array([3., 4.]).astype(np.float32), np.array([5., 6.]).astype(np.float32)]
         expect(node, inputs=[input], outputs=[y for y in expected_outputs], name='test_split_equal_parts_1d')
 
         node = onnx.helper.make_node(
@@ -34,13 +34,13 @@ class Split(Base):
             split=[2, 4]
         )
 
-        expected_outputs = [np.array([1., 2.]), np.array([3., 4., 5., 6.])]
+        expected_outputs = [np.array([1., 2.]).astype(np.float32), np.array([3., 4., 5., 6.]).astype(np.float32)]
         expect(node, inputs=[input], outputs=[y for y in expected_outputs], name='test_split_variable_parts_1d')
 
     @staticmethod
     def export_2d():
         input = np.array([[1., 2., 3., 4., 5., 6.],
-                          [7., 8., 9., 10., 11., 12.]])
+                          [7., 8., 9., 10., 11., 12.]]).astype(np.float32)
 
         node = onnx.helper.make_node(
             'Split',
@@ -49,8 +49,8 @@ class Split(Base):
             axis=1
         )
 
-        expected_outputs = [np.array([[1., 2., 3.], [7., 8., 9.]]),
-                            np.array([[4., 5., 6.], [10., 11., 12.]])]
+        expected_outputs = [np.array([[1., 2., 3.], [7., 8., 9.]]).astype(np.float32),
+                            np.array([[4., 5., 6.], [10., 11., 12.]]).astype(np.float32)]
 
         expect(node, inputs=[input], outputs=[y for y in expected_outputs], name='test_split_equal_parts_2d')
 
@@ -62,7 +62,7 @@ class Split(Base):
             split=[2, 4]
         )
 
-        expected_outputs = [np.array([[1., 2.], [7., 8.]]),
-                            np.array([[3., 4., 5., 6.], [9., 10., 11., 12.]])]
+        expected_outputs = [np.array([[1., 2.], [7., 8.]]).astype(np.float32),
+                            np.array([[3., 4., 5., 6.], [9., 10., 11., 12.]]).astype(np.float32)]
 
         expect(node, inputs=[input], outputs=[y for y in expected_outputs], name='test_split_variable_parts_2d')
